[WIP] Port Collaboration to CollabWrapper
Activity uses telepathy tubes chan, this port aims to move collaboration to collabwrapper.
When a user opens the activity, the images shown in the activity are saved in a zip file and shared to other users when the activity is shared.
The joined users receive the zip file if they have enough disk space for it.

Replace the use of tempfile with activity_zip to save images as a zipfile.
Activity currently takes a while to load as load_journal_table searches for images from the root directory and it's subdirectories.

TODO
- Test collaboration.
- Fix issues with zipfile.
- Restrict image search to journal objects.

Signed-off-by: Ibiam Chihurumnaya <ibiamchihurumnaya@gmail.com>